backend/pyproject.toml
[project]
name = "flocus-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrei Sorokin",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "odmantic",
    "pytest",
    "pydantic[email]",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "python-multipart (>=0.0.20,<0.0.21)",
    "async_generator (>=1.10,<2.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "black>=24.2.0",
    "flake8>=7.0.0",
    "isort>=5.13.0"
]
